fix a bunch of problems in ParEval

+++ drivers/cpp/benchmarks/scan/32_scan_sum_of_prefix_sum_array/baseline.hpp
@@ -9,8 +9,8 @@
    input: [-7, 2, 1, 9, 4, 8]
    output: 15
 */
-double NO_INLINE correctSumOfPrefixSum(std::vector<double> const& x) {
-    std::vector<double> prefixSum(x.size());
+int64_t NO_INLINE correctSumOfPrefixSum(std::vector<int64_t> const& x) {
+    std::vector<int64_t> prefixSum(x.size());
     std::inclusive_scan(x.begin(), x.end(), prefixSum.begin());
-    return std::accumulate(prefixSum.begin(), prefixSum.end(), 0.0);
+    return std::accumulate(prefixSum.begin(), prefixSum.end(), 0);
 }
